requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
bcrypt
pandas
numpy
scikit-learn
joblib
matplotlibfastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
bcrypt
pandas
numpy
scikit-learn
joblib
matplotlib